[VB-25] fix: восстановил переходы на auth и main модули
