--- VolleyBolley/App/LaunchScreen/en.lproj/LaunchScreen.strings
@@ -1,3 +1,0 @@
-
-/* Class = "UILabel"; text = "VOLLEYBOLLEY"; ObjectID = "WRz-4j-RMl"; */
-"WRz-4j-RMl.text" = "VOLLEYBOLLEY";
