+++ VolleyBolley/App/Router/AppRouter.swift
@@ -34,27 +34,48 @@ final class AppRouter {
     // MARK: - Public Methods
 
     func start() {
-        // TODO: Переписать через userSessionService
-        if UserDefaults.standard.isOnboardingShown {
-            showAuthorization()
-        } else {
+        if !userSessionService.isOnboardingShown {
             showOnboarding()
+        } else if !userSessionService.isAuthorized {
+            showAuth()
+        } else {
+            showMainApp()
         }
     }
 
     // MARK: - Private Methods
 
     private func showOnboarding() {
-        guard let onboardingVC = resolver.resolve(OnboardingViewController.self) else {
-            fatalError("OnboardingViewController не зарегистрирован")
+        guard var router = resolver.resolve(OnboardingRouterProtocol.self) else {
+            print("Error: Failed to resolve OnboardingRouterProtocol")
+            return
         }
-        let nav = UINavigationController(rootViewController: onboardingVC)
-        window.rootViewController = nav
-        window.makeKeyAndVisible()
+
+        router.onFinish = { [weak self] in
+            guard let self else { return }
+
+            self.start()
+        }
+
+        onboardingRouter = router
+        window.rootViewController = router.start()
     }
 
-    private func showAuthorization() {
-		// TODO: Добавить переход к авторизации
+    private func showAuth() {
+        guard var router = resolver.resolve(AuthRouterProtocol.self) else {
+            print("Error: Failed to resolve AuthRouterProtocol")
+            return
+        }
+
+        router.onLoginSuccess = { [weak self] in
+            guard let self else { return }
+
+            self.userSessionService.markUserAuthorized()
+            self.start()
+        }
+
+        authRouter = router
+        window.rootViewController = router.start()
     }
 
     private func showMainApp() {
