+++ VolleyBolley/Modules/Onboarding/Interactor/OnboardingInteractorProtocol.swift
@@ -1,0 +1,11 @@
+//
+//  OnboardingInteractorProtocol.swift
+//  VolleyBolley
+//
+//  Created by Nikolai Eremenko on 02.08.2025.
+//
+
+protocol OnboardingInteractorProtocol: AnyObject {
+    func markOnboardingShown()
+//    func getWelcomeItem() -> OnboardingItem
+}
