+++ VolleyBolley/Modules/Onboarding/Presenter/OnboardingPresenterProtocol.swift
@@ -1,0 +1,11 @@
+//
+//  OnboardingPresenterProtocol.swift
+//  VolleyBolley
+//
+//  Created by Nikolai Eremenko on 02.08.2025.
+//
+
+protocol OnboardingPresenterProtocol: AnyObject {
+    func viewDidLoad()
+    func didTapGetStartedButton()
+}
