+++ VolleyBolley/Modules/Onboarding/View/OnboardingViewProtocol.swift
@@ -1,0 +1,10 @@
+//
+//  OnboardingViewProtocol.swift
+//  VolleyBolley
+//
+//  Created by Nikolai Eremenko on 02.08.2025.
+//
+
+protocol OnboardingViewProtocol: AnyObject {
+    //    func configureView(with entity: OnboardingPage)
+}
